Crate flowc

flowc the the "flow compiler" that takes a hierarchical description of flows and functions and compiles it into a network of functions in a Manifest file for execution by flowr or other flow runtimes.

Execute flowc or flowc --help or flowc -h at the comment line for a description of the command line options.

Modules

flow_compile

lib_build

source_arg

Structs

Error
The Error type.

Options

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

get_matches

main

parse_args

run

set_lib_search_path

Type Definitions

Result
Convenient wrapper around std::Result.