Struct flowc::Options

pub struct Options {

    pub(crate) lib: bool,
    pub(crate) source_url: Url,
    pub(crate) flow_args: Vec<String>,
    pub(crate) tables_dump: bool,
    pub(crate) graphs: bool,
    pub(crate) execution_metrics: bool,
    pub(crate) wasm_execution: bool,
    pub(crate) compile_only: bool,
    pub(crate) debug_symbols: bool,
    pub(crate) provided_implementations: bool,
    pub(crate) output_dir: PathBuf,
    pub(crate) stdin_file: Option<String>,
    pub(crate) lib_dirs: Vec<String>,
    pub(crate) native_only: bool,
    pub(crate) context_root: Option<PathBuf>,
    pub(crate) verbosity: Option<String>,
    pub(crate) optimize: bool,

}

Options struct gathers information from the parsing of the command line options to be used to configure execution

Fields
lib: boolsource_url: Urlflow_args: Vec<String>tables_dump: boolgraphs: boolexecution_metrics: boolwasm_execution: boolcompile_only: booldebug_symbols: boolprovided_implementations: booloutput_dir: PathBufstdin_file: Option<String>lib_dirs: Vec<String>native_only: boolcontext_root: Option<PathBuf>verbosity: Option<String>optimize: bool